Remove AdministrationAccessControls and dead SQL access control methods
Removes the SQL-based access control infrastructure from the administration
repository now that service-layer authorization has been migrated to OpenFGA.

- Inline `buildAdministrationUserAssignmentsQuery` and
  `getAssignedUserCountsByAdministrationIds` directly onto
  `AdministrationRepository` (still used by the service for the stats embed)
- Delete dead methods `getAuthorizedById` and `getUserRolesForAdministration`
  (replaced by FGA `requirePermission` in the service layer)
- Delete `AdministrationAccessControls` class and its integration test
- Delete the `createMockAdministrationAccessControls` test support mock

Part of issue #1764.
