+++ apps/backend/src/test-support/repositories/administration.repository.ts
@@ -12,12 +12,10 @@ export function createMockAdministrationRepository(): MockedObject<Administratio
   return {
     ...createMockBaseRepositoryMethods(),
     listAll: vi.fn(),
-    getAuthorizedById: vi.fn(),
     getAssignedUserCountsByAdministrationIds: vi.fn(),
     getDistrictsByAdministrationId: vi.fn(),
     getSchoolsByAdministrationId: vi.fn(),
     getClassesByAdministrationId: vi.fn(),
-    getUserRolesForAdministration: vi.fn(),
     getGroupsByAdministrationId: vi.fn(),
     getTaskVariantsByAdministrationId: vi.fn(),
     getAgreementsByAdministrationId: vi.fn(),
